requirements.txt
aiosmtplib==2.0.2
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.8.1
bcrypt==3.2.0
blinker==1.7.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
fastapi==0.70.0
greenlet==3.0.3
h11==0.14.0
idna==3.6
Jinja2==3.1.3
jose==1.0.0
MarkupSafe==2.1.5
motor==3.4.0
passlib==1.7.4
pyasn1==0.6.0
pycparser==2.22
pydantic==1.10.11
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyJWT==2.3.0
pymongo==4.6.3
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.32
starlette==0.16.0
typing_extensions==4.10.0
uvicorn==0.15.0
